+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+## 0.1.2
+
+* (01) Removes `--email` and `--password`
+* (02) Supports downloading "references" of a course
+
 ## 0.1.1
 
-* Add `--cookies`
+* (01) Adds `--cookies`
 
 ## 0.1.0